library export: Rework `LibraryExporter` etc.
Instead of `LibraryExporter` having a `LibraryExportModel`,
`DlgLibraryExport` now produces a `LibraryExportModel`.

In this workflow, `LibraryExportModel` is passed by value instead
of by reference. Its contained attributes are implicitly shared by
Qt anyways.

With these changes, the `LiraryExportModel::clear()` method is no
longer used and was therefore removed. If desired in the future,
one can clear a `LibraryExportModel` by simply assigning a fresh
one to it.

Attempts to start multiple sessions are handled as follows.

* `DlgLibraryExport`

  Instead of opening a new dialog, the existent one is activated.
  This holds true, even if it was closed in the meantime.

* `LibraryExportWorker`

  Simply nothing happens when trying to start an export worker
  while one is active (at least if the attempts originate from
  the same `DlgLibraryExport`. As mentioned above, there's only
  one of them). In order to implement this, the
  `LibraryExporter` hat to subscribe itself to all three ending
  signals with which the `LibraryExportWorker` can finish its
  work.

  TODO: Instead of doing nothing, activate the worker's window.