Added track analysis & waveform export.
Tracks are now scheduled for analysis during a library export, ensuring
that a beat grid and waveform will be available.

The calls to the analysis engine require the job to be able to emit
signals and receive slot calls, meaning it must be a QObject.  As such,
the prior job scheduler has been removed and replaced with a simple
QThread-based implementation.